pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "camelot-py[cv]>=1.0.0",
    "email-validator>=2.2.0",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "gunicorn>=23.0.0",
    "numpy>=2.3.1",
    "opencv-python>=4.11.0.86",
    "openpyxl>=3.1.5",
    "pandas>=2.3.1",
    "pdfplumber>=0.11.7",
    "pillow>=11.3.0",
    "psycopg2-binary>=2.9.10",
    "pytesseract>=0.3.13",
    "python-docx>=1.2.0",
    "tabula-py>=2.10.0",
    "werkzeug>=3.1.3",
]
